Add files via upload
In this project I have web scrape the data of used certified cars by using requests and beautifusoup libraries and dump this data in excel sheet for further analysis for machine learning. 
ETL is one of the most important tool to extract the data from third party websites, api's and to perform some actions so that the data is in well structured format (transform) and loaded (dumped) that data in sql for further usage.