Add files via upload
The objective of this project is to accurately predict insurance costs based on people’s data, including age, Body Mass Index, smoking or not, etc. Additionally, I will also determine what the most important variable influencing insurance costs is. These estimates could be used to create actuarial tables that set the price of yearly premiums higher or lower according to the expected treatment costs. This is a regression problem statement.